+++ ORStackViewExample/ORSecondViewController.m
@@ -12,31 +12,14 @@
 
 // More complex use case: Inserting 
 
-@interface ORSecondViewController ()
-@property (nonatomic, strong) ORStackView *stackView;
-@end
-
 @implementation ORSecondViewController
 
 - (UIStatusBarStyle)preferredStatusBarStyle {
     return UIStatusBarStyleLightContent;
 }
 
-- (void)loadView
-{
-    self.view = [[UIView alloc] init];
-}
-
 - (void)viewDidLoad
 {
-    self.stackView = [[ORStackView alloc] init];
-    [self.view addSubview:self.stackView];
-    if ([self respondsToSelector:@selector(topLayoutGuide)]) {
-        [self.view addConstraint:[NSLayoutConstraint constraintWithItem:self.stackView attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self.topLayoutGuide attribute:NSLayoutAttributeBottom multiplier:1.0 constant:0]];
-    }
-    [self.view addConstraint:[NSLayoutConstraint constraintWithItem:self.stackView attribute:NSLayoutAttributeWidth relatedBy:NSLayoutRelationEqual toItem:self.view attribute:NSLayoutAttributeWidth multiplier:1.0 constant:0]];
-    [self.view addConstraint:[NSLayoutConstraint constraintWithItem:self.stackView attribute:NSLayoutAttributeCenterX relatedBy:NSLayoutRelationEqual toItem:self.view attribute:NSLayoutAttributeCenterX multiplier:1.0 constant:0]];
-
     ORColourView *view1 = [[ORColourView alloc] init];
     view1.text = @"1 - ORStackView - Tap Me";
     view1.fakeContentSize = (CGSize){ UIViewNoIntrinsicMetric , 40};
@@ -54,11 +37,11 @@
     ORColourView *view4 = [[ORColourView alloc] init];
     view4.text = @"4 - Lorem ipsum, etc. etc.";
     view4.fakeContentSize = (CGSize){ UIViewNoIntrinsicMetric , 20 };
-
-    [self.stackView insertSubview:view2 atIndex:0 withTopMargin:20.0 sideMargin:20.0];
-    [self.stackView insertSubview:view4 atIndex:1 withTopMargin:15.0 sideMargin:20.0];
-    [self.stackView insertSubview:view1 atIndex:0 withTopMargin:10.0 sideMargin:20.0];
-    [self.stackView insertSubview:view3 atIndex:2 withTopMargin:10.0 sideMargin:20.0];
+    
+    [self.stackView insertSubview:view2 atIndex:0 withPrecedingMargin:20.0 sideMargin:20.0];
+    [self.stackView insertSubview:view4 atIndex:1 withPrecedingMargin:15.0 sideMargin:20.0];
+    [self.stackView insertSubview:view1 atIndex:0 withPrecedingMargin:10.0 sideMargin:20.0];
+    [self.stackView insertSubview:view3 atIndex:2 withPrecedingMargin:10.0 sideMargin:20.0];
 }
 
 - (void)addView
@@ -66,8 +49,8 @@
     ORColourView *view = [[ORColourView alloc] init];
     view.text = @"new view added at index 2\nTap to remove";
     view.fakeContentSize = (CGSize){ UIViewNoIntrinsicMetric , 50 };
-
-    [self.stackView insertSubview:view atIndex:2 withTopMargin:10.0 sideMargin:10.0];
+    
+    [self.stackView insertSubview:view atIndex:2 withPrecedingMargin:10.0 sideMargin:10.0];
     UITapGestureRecognizer *tapGesture = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(removeTappedView:)];
     [view addGestureRecognizer:tapGesture];
 }
@@ -76,4 +59,5 @@
 {
     [self.stackView removeSubview:gesture.view];
 }
+
 @end